Moving scrubber to separate component
